requirements.txt
streamlit
fastai
torch
torchvision
Pillow